Use bitvector that we initialize in __init__
Since we already know how many vertices that we have in our FloorGraph __init__ we can initialize our bitvector there causing the time and space complexity to be O(|V| + |E| +|V|)= O(|V| + |E|). Then everytime we remove a node from the priority queue we use this bitvector to check if the current vertex is an exit vertex which has time complexity O(1). We know that we can use bitvector because the id of our location is always an integer.