added djikstra with codntions
The graph where I have a scenario here about a python graph problem that uses Djikstra's. In this graph your goal is to find the smallest total weight needed to get from entrance to exit and the sequence of vertices to go there. However, there is a special condition on when you can end the algortihm. Some of the vertices in the graph has weight, you can choose to add the weight or not to add the weight. But to end the algortihm you need to have added at least a weight from one of the weighted vertices and is currently in that exit vertex.